1.3
- Added preparations for unlimited rewards
- Fixed some bugs
- Reformated some code
- Decided to complete the plugin!
- PAPI support

+++ src/main/java/me/dailyreward/Commands/CommandList/RewardPurge.java
@@ -3,6 +3,7 @@ package me.dailyreward.Commands.CommandList;
 import me.dailyreward.Commands.BaseCommand;
 import me.dailyreward.Commands.Command;
 import me.dailyreward.Commands.CommandArgs;
+import me.dailyreward.Configuration.Config;
 import me.dailyreward.DailyReward;
 import me.dailyreward.Databases.MySQL.PlayerMySQL;
 import me.dailyreward.Utils.Color;
@@ -14,22 +15,23 @@ public class RewardPurge extends BaseCommand {
 
 	private final DailyReward plugin = DailyReward.getInstance();
 	private final PlayerMySQL playerData = new PlayerMySQL();
+	private final Config config = DailyReward.getInstance().getConfiguration();
+
 
 	@Override
 	@Command(name = "reward.purge", isAdminOnly = true, isGameOnly = false)
 	public void onCommand(CommandArgs args) throws SQLException {
 		Player player = args.getPlayer();
 
-		//TODO check for args
+		if ( args.length() > 1 ) {
+			Color.sendMessage(config.getString("MESSAGE_ARGS_ERROR").replace("%command%", "/reward reload"), args.getSender());
+			return;
+		}
 
-		switch (args.getArgs(0).toLowerCase()) {
-			case "mongodb":
-				break;
-			case "mysql":
-				playerData.dropTable();
-				break;
-			default:
-				Color.sendMessage("MESSAGE_ARGS_PURGE", player);
+		if ( "mysql".equals(args.getArgs(0) ) ) {
+			playerData.dropTable();
+		} else {
+			Color.sendMessage("MESSAGE_ARGS_PURGE", player);
 		}
 
 		Color.sendMessage(plugin.getPrefix() + plugin.getConfiguration().getString("MESSAGE_SUCCESS_PURGE"), player);
